feat(ui): refactor utility imports
- Updated imports in various components and routes to use the new utility paths for better organization.
- Introduced new utility functions for color value retrieval and markdown parsing, enhancing the UI's theming capabilities and content rendering.
- Added a scroll spy utility to improve user experience by tracking active sections during scrolling.

These changes streamline the codebase and improve the overall functionality of the UI components.
